--- library/src/main/res/drawable/control_color.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:state_enabled="false" android:color="@color/md_grey_400"/>
-    <item android:color="@color/colorPrimary"/>
-</selector>
