+++ library/src/main/res/drawable/ic_check_white_24dp.xml
@@ -1,5 +1,5 @@
-<vector android:height="24dp" android:tint="#FFFFFF"
+<vector android:height="24dp"
     android:viewportHeight="24.0" android:viewportWidth="24.0"
     android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#FF000000" android:pathData="M9,16.17L4.83,12l-1.42,1.41L9,19 21,7l-1.41,-1.41z"/>
+    <path android:fillColor="#FFF" android:pathData="M9,16.17L4.83,12l-1.42,1.41L9,19 21,7l-1.41,-1.41z"/>
 </vector>
