--- library/src/main/res/layout/material_loading.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.v7.widget.AppCompatEditText xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/edtTxtSpinner"
-    style="@style/MaterialSpinnerStyle"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:clickable="false"
-    android:cursorVisible="false"
-    android:focusable="false"
-    android:focusableInTouchMode="false"
-    android:saveEnabled="false"
-    android:textIsSelectable="false" />
