+++ library/src/main/res/values/attrs.xml
@@ -1,25 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-
-    <attr name="listSelectorWidgetStyle" format="reference" />
-
     <declare-styleable name="ListSelectorView">
         <attr name="android:enabled" />
-        <attr name="android:entries" />
         <attr name="android:textColor" />
-
-
         <attr name="android:textSize" />
         <attr name="android:drawableEnd"/>
         <attr name="android:hint"/>
         <attr name="lsw_value" format="string"/>
         <attr name="lsw_spaceSize" format="dimension"/>
-    </declare-styleable>
-
-    <declare-styleable name="MaterialSpinner">
-        <attr name="android:enabled" />
-        <attr name="android:entries" />
-        <attr name="android:textColor" />
-        <attr name="colorAccent" format="color" />
     </declare-styleable>
 </resources>
