+++ library/src/main/res/values/colors.xml
@@ -1,7 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <color name="hint_color">#808080</color>
-    <color name="md_grey_400">#BDBDBD</color>
     <color name="window_background">@android:color/white</color>
     <color name="colorPrimary">#0af</color>
 </resources>
