+++ library/src/main/res/values/styles.xml
@@ -1,23 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <style name="Dialog"/>
-
-    <style name="ListSelectorWidget">
-        <item name="android:textColor">@color/colorPrimary</item>
-        <item name="android:theme">@style/MaterialSpinnerTheme</item>
-        <item name="android:paddingTop">8dp</item>
-        <item name="android:clickable">true</item>
-    </style>
-
-    <style name="MaterialSpinnerStyle" parent="Widget.AppCompat.Spinner.Underlined">
-        <item name="android:textColor">@drawable/control_color</item>
-        <item name="android:theme">@style/MaterialSpinnerTheme</item>
-        <item name="android:paddingTop">8dp</item>
-    </style>
-
-    <style name="MaterialSpinnerTheme">
-        <item name="colorControlNormal">@drawable/control_color</item>
-        <item name="colorControlActivated">@drawable/control_color</item>
-        <item name="colorControlHighlight">@drawable/control_color</item>
-    </style>
 </resources>
