+++ sample/src/main/res/layout/activity_main.xml
@@ -9,14 +9,6 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingBottom="@dimen/activity_vertical_margin">
 
-    <com.github.jura73.materialspinner.ListOneSelectorView
-        android:id="@+id/listSelector1"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:entries="@array/list_of_cities_in_Italy"
-        android:hint="Производитель"
-        app:lsw_value="Производитель"/>
-
     <com.github.jura73.materialspinner.ListMultiSelectorView
         android:id="@+id/listSelector2"
         android:layout_width="match_parent"
@@ -30,7 +22,6 @@
         android:layout_height="wrap_content"
         android:entries="@array/list_of_cities_in_Italy"
         android:hint="cities in Italy"
-        android:background="@color/md_grey_400"
         app:passwordToggleDrawable="@android:drawable/arrow_up_float" />
 
     <com.github.jura73.materialspinner.MaterialSpinner
